+++ R/body.R
@@ -13,7 +13,9 @@ check_message_body <- function(content) {
 #' @param disposition How content is presented (Content-Disposition).
 #' @param charset How content is encoded.
 #' @param encoding How content is transformed to ASCII (Content-Transfer-Encoding).
-#' @param interpolate Whether to interpolate using \code{glue}.
+#' @param interpolate Whether or not to interpolate into input using \link[glue]{glue}.
+#' @param .open The opening delimiter.
+#' @param .close The closing delimiter.
 #' @param .envir Environment used for \code{glue} interpolation. Defaults to \code{parent.frame()}.
 #'
 #' @return A message object.
@@ -42,32 +44,27 @@ text <- function(
   charset = "utf-8",
   encoding = "7bit",
   interpolate = TRUE,
+  .open = "{{",
+  .close = "}}",
   .envir = NULL
 ) {
+  check_message_body(content)
+
   if (is.null(.envir)) .envir = parent.frame()
   else .envir = list2env(.envir)
 
-  check_message_body(content)
+  if (interpolate) content <- glue(content, .open = .open, .close = .close, .envir = .envir)
 
-  if (interpolate) content <- glue(content, .envir = .envir)
+  body <- text_plain(content, disposition, charset, encoding)
 
-  type <- "text/plain"
-
-  body <- mime(type, disposition, charset, encoding, NA)
-  body$body <- content
-
-  msg$parts <- c(msg$parts, list(body))
+  msg <- append(msg, body)
 
   if (get_option_invisible()) invisible(msg) else msg
 }
 
 #' Add an HTML body to a message object.
 #'
-#' @param msg A message object.
-#' @param content A string of message content or the path to a file containing HTML.
-#' @param disposition How content is presented (Content-Disposition).
-#' @param charset How content is encoded.
-#' @param encoding How content is transformed to ASCII (Content-Transfer-Encoding).
+#' @inheritParams text
 #' @return A message object.
 #' @seealso \code{\link{text}}
 #' @export
@@ -79,7 +76,17 @@ text <- function(
 #'
 #' # Read HTML message from a file.
 #' msg <- envelope() %>% html("message.html")
-html <- function(msg, content, disposition = "inline", charset = "utf-8", encoding = "quoted-printable") {
+html <- function(
+  msg,
+  content,
+  disposition = "inline",
+  charset = "utf-8",
+  encoding = "quoted-printable",
+  interpolate = TRUE,
+  .open = "{{",
+  .close = "}}",
+  .envir = NULL
+) {
   check_message_body(content)
 
   # Check if it's a file.
@@ -88,12 +95,14 @@ html <- function(msg, content, disposition = "inline", charset = "utf-8", encodi
     content <- paste(readLines(content), collapse = "\n")
   }
 
-  type <- "text/html"
+  if (is.null(.envir)) .envir = parent.frame()
+  else .envir = list2env(.envir)
 
-  body <- mime(type, disposition, charset, encoding, NA)
-  body$body <- qp_encode(content)
+  if (interpolate) content <- glue(content, .open = .open, .close = .close, .envir = .envir)
 
-  msg$parts <- c(msg$parts, list(body))
+  body <- text_html(content, disposition, charset, encoding)
 
-  invisible(msg)
+  msg <- append(msg, body)
+
+  if (get_option_invisible()) invisible(msg) else msg # nocov
 }