+++ R/mime.R
@@ -1,99 +1,295 @@
-#' Create a MIME (Multipurpose Internet Mail Extensions) object.
+ERROR_NOT_MIME_OBJECT <- "Child is not a MIME object."
+
+# CONSTRUCTOR -----------------------------------------------------------------
+
+is.mime <- function(x) {
+  "MIME" %in% class(x)
+}
+
+#' Create a MIME object
 #'
-#' @param content_type The MIME type of the content.
-#' @param content_disposition Should the content be displayed inline or as an attachment?
-#' @param encoding How to encode binary data to ASCII.
+#' ```
+#' MIME
+#'  ├── multipart/mixed
+#'  ├── multipart/related
+#'  ├── text/plain
+#'  ├── text/html
+#'  └── other
+#' ```
+#'
+#' @section MIME Multipart Types:
+#'
+#' There are a number of options for multipart messages:
+#'
+#' \itemize{
+#'  \item{\code{multipart/mixed} — }{Used for sending content with multiple independent parts either inline or as attachments. Each part can have different \code{Content-Type}.}
+#'  \item{\code{multipart/alternative} — }{Used when each part of the message is an "alternative" version of the same content. The order of the parts is important: preferred and/or more complex formats should be found towards the end.
+#'
+#'  \emph{Example:} A message with both plain text and HTML versions.}
+#'  \item{\code{multipart/digest} — }{Used to send multiple plain text messages.}
+#'  \item{\code{multipart/related} — }{Used when each part of the the message represents a component of the complete message.
+#'
+#'  \emph{Example:} A web page with images.}
+#'  \item{\code{multipart/signed} — }{Used when a message has a digital signature attached.}
+#'  \item{\code{multipart/encrypted} — }{Used for a message with encrypted content.}
+#' }
+#'
+#' A nice illustration of how some of these relate can be found at \url{https://stackoverflow.com/a/40420648/633961}.
+#'
+#' @noRd
+#'
+#' @param content Content.
+#' @param disposition Should the content be displayed inline or as an attachment? Valid options are \code{"inline"} and \code{"attachment"}. If set to \code{NA} then will guess appropriate value.
 #' @param charset How to interpret the characters in the content. Most often either UTF-8 or ISO-8859-1.
-#' @param cid An optional Content-Id.
-#' @param ... Other arguments.
+#' @param encoding How to encode binary data to ASCII.
+#' @param boundary Boundary string.
+#' @param type The MIME type of the content.
+#' @param children
+#'
 #' @return A MIME object.
-mime <- function(content_type, content_disposition, charset, encoding, cid = NA, ...) {
+MIME <- function(
+  content = NULL,
+  disposition = NA,
+  charset = NA,
+  encoding = NA,
+  boundary = emayili:::hexkey(),
+  type = NA,
+  children = list()
+) {
+  # If just a single child, convert to list.
+  if (!all(class(children) == c("list"))) children <- list(children)
+  # Check that all children are MIME.
+  for (child in children) {
+    if (!is.mime(child)) stop(ERROR_NOT_MIME_OBJECT, call. = FALSE)
+  }
+
   structure(
     list(
-      header = list(
-        content_type = content_type,
-        content_disposition = content_disposition,
-        encoding = encoding,
-        format = format,
-        charset = charset,
-        cid = cid,
+      content = content,
+      disposition = disposition,
+      charset = charset,
+      encoding = encoding,
+      boundary = boundary,
+      children = children,
+      type = type
+    ),
+    class = "MIME"
+  )
+}
+
+multipart_related <- function(...) {
+  structure(
+    c(
+      MIME(...),
+      list()
+    ),
+    class = c("multipart_related", "MIME")
+  )
+}
+
+multipart_mixed <- function(...) {
+  structure(
+    c(
+      MIME(...),
+      list()
+    ),
+    class = c("multipart_mixed", "MIME")
+  )
+}
+
+text_plain <- function(
+  content,
+  disposition = "inline",
+  charset = "utf-8",
+  encoding = "7bit",
+  ...
+) {
+  structure(
+    c(
+      MIME(content, disposition, charset, encoding, boundary = NA, ...),
+      list()
+    ),
+    class = c("text_plain", "MIME")
+  )
+}
+
+#' Create \code{text/html} MIME object
+#'
+#' @noRd
+#'
+#' @inheritParams MIME
+#' @param squish Whether to remove whitespace outside of tags.
+#' @param ... Further arguments passed to or from other methods.
+#'
+#' @return A \code{text_html} object derived from \code{MIME}.
+text_html <- function(
+  content,
+  disposition = "inline",
+  charset = "utf-8",
+  encoding = NA,
+  squish = FALSE,
+  ...
+) {
+  # Clean up content.
+  #
+  if (squish) {
+    content <- content %>%
+      str_replace_all("(^|(?<=>))[:space:]+($|(?=<))", "")
+  }
+
+  # Remove empty lines.
+  content <- content %>% str_replace_all("(\n)+", "\n")
+
+  structure(
+    c(
+      MIME(
+        content,
+        disposition,
+        charset,
+        encoding,
+        boundary = NA,
         ...
       ),
-      body = NULL
+      list()
     ),
-    class="mime")
+    class = c("text_html", "MIME")
+  )
 }
 
-#' Encode parameter values for MIME headers
+#' Other (not text or HTML) MIME content
 #'
-#' Formats UTF-8 parameter values for the use in MIME headers. Non-ASCII
-#' characters, control characters, and some special characters have to be
-#' specially encoded according to the
-#' \href{https://tools.ietf.org/html/rfc2231}{RFC2231} specification.
+#' @noRd
 #'
-#' @param x \code{character(1)}. UTF-8 string to format.
-#' @return \code{character(1)}. String to put right after the parameter name in
-#'   a MIME header. The equal sign and possible quotation marks are included.
-#'   For example \code{"I'm not quoted.csv"} turns to
-#'   \code{"=\\"I'm not quoted.csv\\""} while \code{"\\"I'm quoted\\".csv"}
-#'   results in \code{"*=utf-8''\%22I'm\%20quoted\%22.csv"}.
-#' @examples
-#' emayili:::parameter_value_encode("I'm not quoted.csv")
-#' emayili:::parameter_value_encode("\"I'm quoted\".csv")
-parameter_value_encode <- function(x){
-  x_raw <- charToRaw(x)
-  ascii_to_encode <- as.raw(c(0x00:0x1F, 0x22, 0x5C, 0x7F))
-  # testing showed that control characters, "\"" and sometimes "\\" also must be
-  # encoded
-  ascii <- x_raw <= as.raw(0x7F) & !(x_raw %in% ascii_to_encode)
-  if (all(ascii)) {
-    return(paste0("=\"", x, "\""))
+#' @inheritParams MIME
+#'
+#' @param filename Path to a file.
+#' @param name Name used when downloading file.
+#' @param cid An optional Content-Id.
+#' @param ... Further arguments passed to or from other methods.
+other <- function(
+  filename,
+  name = NA,
+  type = NA,
+  cid = NA,
+  disposition = "attachment",
+  encoding = "base64",
+  ...
+) {
+  charset <- NA
+  basename <- basename(filename)
+  name <- ifelse(is.na(name), basename, name)
+
+  if (!is.na(type)) {
+    # Could use mime::mimemap to map from specific extensions to MIME types,
+    # so that the following would give the same result:
+    #
+    # attachment(..., type = "pdf")
+    # attachment(..., type = "application/pdf")
   } else {
-    syntax_ascii <- as.raw(c(0x20, 0x25, 0x3B))
-    # in case of encoding, " ", "%" and ";" must be encoded as well
-    ascii <- ascii & !(x_raw %in% syntax_ascii)
-    encoded <- character(length(x_raw))
-    encoded[ascii] <- rawToChar(x_raw[ascii], multiple = TRUE)
-    encoded[!ascii] <- paste0("%", toupper(as.character(x_raw[!ascii])))
-    return(paste0("*=utf-8''", paste(encoded, collapse = "")))
+    type <- guess_type(filename, empty = "application/octet-stream")
   }
+  type <- glue('{type}; name="{name}"')
+
+  if (is.na(disposition)) {
+    disposition <- ifelse(
+      grepl("text", type),
+      # If it's text...
+      "inline",
+      # ... otherwise:
+      "attachment"
+    )
+  }
+  disposition <- glue('{disposition}; filename="{basename}"')
+
+  content <- base64encode(
+    read_bin(filename),
+    76L,
+    "\r\n"
+  )
+
+  structure(
+    c(
+      MIME(content, disposition, charset, encoding, boundary = NA, type = type, ...),
+      list(
+        cid = ifelse(is.na(cid), hexkey(), cid)
+      )
+    ),
+    class = c("attachment", "MIME")
+  )
 }
 
-#' Format the header of a MIME object.
+# APPEND ----------------------------------------------------------------------
+
+#' Append child to a MIME element
 #'
-#' The header must conform to the SMTP Protocol (https://tools.ietf.org/html/rfc821). Specifically, lines should be
-#' terminated by CRLF (not just LF). The \dQuote{name} and \dQuote{filename}
-#' fields can contain any possible UTF-8 characters as they are
-#' \link[=parameter_value_encode]{parameter value encoded}.
+#' This is a generic function.
 #'
-#' @param msg A message object.
-#' @return A formatted header string.
-format.mime <- function(msg) {
-  # Unpack relevant variables here so that we can contain the environment which
-  # glue() searches.
+#' @param x MIME element
+#' @param child Child MIME element
+append <- function(x, child) {
+  UseMethod("append", x)
+}
+append.MIME <- function(x, child) {
+  if (!is.mime(child)) stop(ERROR_NOT_MIME_OBJECT, call. = FALSE)
+  x$children <- c(x$children, list(child))
+  x
+}
+
+# CHARACTER -------------------------------------------------------------------
+
+#' Convert MIME object to character vector
+#'
+#' @param x MIME object
+#' @param ... Further arguments passed to or from other methods.
+as.character.MIME <- function(x, ...) {
+  children <- sapply(x$children, function(child) {
+    paste(paste0("--", x$boundary), as.character.MIME(child), sep = "\n")
+  })
+  type <- ifelse(!is.na(x$type), x$type, sub("_", "/", class(x)[1]))
   #
-  content_type        <- msg$header$content_type
-  charset             <- msg$header$charset
-  name                <- msg$header$name
-  content_disposition <- msg$header$content_disposition
-  filename            <- msg$header$filename
-  cid                 <- msg$header$cid
-  encoding            <- msg$header$encoding
+  body <- c(
+    # Head.
+    paste(
+      c(
+        glue('Content-Type:              {type}'),
+        if (!is.na(x$charset)) glue('charset={x$charset}') else NULL,
+        if (!is.na(x$boundary)) glue('boundary="{x$boundary}"') else NULL
+      ),
+      collapse = "; "
+    ),
+    if (!is.na(x$disposition)) {
+      glue('Content-Disposition:       {x$disposition}')
+    } else NULL,
+    if (!is.na(x$encoding)) {
+      glue('Content-Transfer-Encoding: {x$encoding}')
+    } else NULL,
+    if (!is.null(x$cid)) {
+      paste(
+        glue('X-Attachment-Id:           {x$cid}'),
+        glue('Content-ID:                <{x$cid}>'),
+        sep = "\n"
+      )
+    } else NULL,
+    "",
+    # Content (if any).
+    x$content,
+    # Children (if any).
+    if(length(children)) children else NULL,
+    # Foot.
+    if (!is.na(x$boundary)) glue('--{x$boundary}--') else NULL
+  )
 
-  if (!is.null(name))     name     <- parameter_value_encode(name)
-  if (!is.null(filename)) filename <- parameter_value_encode(filename)
+  paste(body, collapse = "\n")
+}
 
-  headers <-  c(
-    'Content-Type: {content_type}',
-    ifelse(!is.null(charset), '; charset={charset}', ''),
-    ifelse(!is.null(name), '; name{name}', ''),
-    '\r\nContent-Disposition: {content_disposition}',
-    ifelse(!is.null(filename), '; filename{filename}', ''),
-    ifelse(!is.na(cid), '\r\nContent-Id: <{cid}>\r\nX-Attachment-Id: {cid}', ''),
-    '\r\nContent-Transfer-Encoding: {encoding}'
-  ) %>%
-    paste(collapse = "") %>%
-    glue()
+# PRINT -----------------------------------------------------------------------
 
-  paste(headers, msg$body, sep = "\n\n")
+#' Print a MIME object
+#'
+#' @noRd
+#'
+#' @param x MIME object
+#' @param ... Further arguments passed to or from other methods.
+print.MIME <- function(x, ...) {
+  cat(as.character(x))
 }