+++ R/utils.R
@@ -1,6 +1,6 @@
 #' Pipe operator
 #'
-#' FOO \link[magrittr]{%>%}
+#' \link[magrittr]{%>%}
 #'
 #' @name %>%
 #' @rdname pipe
@@ -35,4 +35,52 @@ get_option_invisible <- function(default = TRUE) {
 
 get_option_details <- function(default = TRUE) {
   getOption("envelope_details", default = default)
+}
+
+#' Read entire text file into character vector
+#'
+#' @noRd
+#'
+#' @param path Relative or absolute file path
+#'
+#' @return A character vector
+read_text <- function(path) {
+  readChar(path, file.info(path)$size)
+}
+
+#' Read entire binary file into character vector
+#'
+#' @noRd
+#'
+#' @param path Relative or absolute file path
+#'
+#' @return A character vector
+read_bin <- function(path) {
+  readBin(path, "raw",  file.info(path)$size)
+}
+
+#' Normalise file path
+#'
+#' @noRd
+#'
+#' @param path Relative or absolute file path
+#'
+#' @return An absolute file path (if the file exists) or \code{NA}.
+normalise_filepath <- function(path) {
+  possibly(normalizePath, NA_character_)(path, mustWork = TRUE)
+}
+
+#' Check if character vector is a file name or file path
+#'
+#' @noRd
+#'
+#' @param path A character vector (which might also be a file path)
+#'
+#' @return If it is a file path, then return \code{TRUE}, otherwise return \code{FALSE}.
+is_filepath <- function(path) {
+  !is.na(normalise_filepath(path))
+}
+
+hexkey <- function(object = runif(1), algorithm="crc32") {
+  digest(object, algorithm)
 }